refactor: move event handlers
